requirements.txt
Flask
PyMySQL
SQLAlchemy
Werkzeug
llama-cloud
llama-index
llama-index-agent-openai
llama-index-cli
llama-index-core
llama-index-embeddings-jinaai
llama-index-embeddings-openai
llama-index-indices-managed-llama-cloud
llama-index-legacy
llama-index-llms-ollama
llama-index-llms-openai
llama-index-multi-modal-llms-openai
llama-index-program-openai
llama-index-question-gen-openai
llama-index-readers-file
llama-index-readers-llama-parse
llama-index-readers-web
llama-index-vector-stores-tidbvector
llama-parse
mysql-connector-python
pillow
python-dotenv
python-pptx
supabase
Gunicorn